Added setting for disabling directory deletion

+++ Program/src/org/jphototagger/program/module/directories/DeleteDirectoryController.java
@@ -4,11 +4,14 @@ import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 import java.io.File;
 import javax.swing.tree.DefaultMutableTreeNode;
+import org.jphototagger.api.preferences.Preferences;
+import org.jphototagger.api.preferences.PreferencesKeys;
 import org.jphototagger.lib.awt.EventQueueUtil;
 import org.jphototagger.lib.io.TreeFileSystemDirectories;
 import org.jphototagger.lib.swing.AllSystemDirectoriesTreeModel;
 import org.jphototagger.program.factory.ModelFactory;
 import org.jphototagger.program.module.filesystem.FileSystemDirectories;
+import org.openide.util.Lookup;
 
 /**
  * Listens to {@code DirectoriesPopupMenu#getItemDeleteDirectory()} and
@@ -37,8 +40,10 @@ public final class DeleteDirectoryController extends DirectoryController {
 
     @Override
     protected void action(final DefaultMutableTreeNode node) {
+        if (!isDeleteDirectoriesEnabled()) {
+            return;
+        }
         File dir = getDirOfNode(node);
-
         if (dir != null) {
             if (FileSystemDirectories.delete(dir)) {
                 EventQueueUtil.invokeInDispatchThread(new Runnable() {
@@ -51,5 +56,12 @@ public final class DeleteDirectoryController extends DirectoryController {
                 });
             }
         }
+    }
+
+    private boolean isDeleteDirectoriesEnabled() {
+        Preferences prefs = Lookup.getDefault().lookup(Preferences.class);
+        return prefs.containsKey(PreferencesKeys.KEY_ENABLE_DELETE_DIRECTORIES)
+                ? prefs.getBoolean(PreferencesKeys.KEY_ENABLE_DELETE_DIRECTORIES)
+                : true;
     }
 }
